% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{create_verhoeff_matrices}
\alias{create_verhoeff_matrices}
\title{create_verhoeff_matrices}
\usage{
create_verhoeff_matrices()
}
\value{
An object with three fields; d5, d5_p, and inv_v
}
\description{
create_verhoeff_matrices
}
\examples{
verhoeff::create_verhoeff_matrices()$d5
verhoeff::create_verhoeff_matrices()$d5_p
verhoeff::create_verhoeff_matrices()$inv_v
}
\keyword{internal}
