% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_d_matrix}
\alias{extract_d_matrix}
\title{extract_d_matrix}
\usage{
extract_d_matrix()
}
\value{
A dataframe
}
\description{
Load the dihedral D5 matrix into memory. One of the matrices used in
check digit calculation
}
\examples{
extract_d_matrix()
}
