% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_f_matrix}
\alias{extract_f_matrix}
\title{extract_f_matrix}
\usage{
extract_f_matrix()
}
\value{
A dataframe
}
\description{
Load the F matrix into memory. One of the matrices used in
check digit calculation
}
\examples{
extract_f_matrix()
}
