% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_inv_matrix}
\alias{extract_inv_matrix}
\title{extract_inv_vector}
\usage{
extract_inv_matrix()
}
\value{
A dataframe
}
\description{
Load the INV matrix into memory. One of the matrices used in
check digit calculation
}
\examples{
extract_inv_matrix()
}
