% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{load_matrices}
\alias{load_matrices}
\title{load_matrices}
\usage{
load_matrices()
}
\value{
Three dataframes
}
\description{
Add all matrices for verhoeff calculation into memory
}
\examples{
load_matrices()
}
