% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verhoeff.R
\name{verhoeff_append}
\alias{verhoeff_append}
\title{verhoeff_append}
\usage{
verhoeff_append(number, sep = "-")
}
\arguments{
\item{number}{The number to calculate a check digit for}

\item{sep}{A separator for the two numbers}
}
\value{
Numeric vector of length equal to its input
}
\description{
Return a number with its check digit appended
}
\examples{
verhoeff::verhoeff_append(123)
}
