% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verhoeff.R
\name{verhoeff_calculate}
\alias{verhoeff_calculate}
\title{verhoeff_calculate}
\usage{
verhoeff_calculate(number, as_list = FALSE)
}
\arguments{
\item{number}{The vector of numbers you want a check digit for}

\item{as_list}{Return the results as a list? Defaults to false}
}
\value{
Vector or list of check digits
}
\description{
verhoeff_calculate
}
\examples{
verhoeff_calculate(1234)
}
